mapper-framework: reposition to SDK
Moved mapper-framework boilerplate code to pkg/mapper/device and pkg/mapper/data
to provide an SDK instead of a template copy. Added driver.Adapter interface.

Fixes #6564
