--- staging/src/github.com/kubeedge/mapper-framework/_template/mapper/Dockerfile_stream
@@ -1,35 +1,0 @@
-FROM golang:1.21.11-bullseye AS builder
-
-WORKDIR /build
-
-ENV GO111MODULE=on \
-    GOPROXY=https://goproxy.cn,direct
-
-COPY . .
-
-RUN apt-get update && \
-    apt-get install -y bzip2 curl upx-ucl gcc-aarch64-linux-gnu libc6-dev-arm64-cross gcc-arm-linux-gnueabi libc6-dev-armel-cross libva-dev libva-drm2 libx11-dev libvdpau-dev libxext-dev libsdl1.2-dev libxcb1-dev libxau-dev libxdmcp-dev yasm
-
-RUN curl -sLO https://ffmpeg.org/releases/ffmpeg-4.1.6.tar.bz2 && \
-    tar -jx --strip-components=1 -f ffmpeg-4.1.6.tar.bz2 &&  \
-    ./configure &&  make && \
-    make install
-
-RUN GOOS=linux go build -o main cmd/main.go
-
-FROM ubuntu:18.04
-
-RUN mkdir -p kubeedge
-
-RUN apt-get update && \
-    apt-get install -y bzip2 curl upx-ucl gcc-aarch64-linux-gnu libc6-dev-arm64-cross gcc-arm-linux-gnueabi libc6-dev-armel-cross libva-dev libva-drm2 libx11-dev libvdpau-dev libxext-dev libsdl1.2-dev libxcb1-dev libxau-dev libxdmcp-dev yasm
-
-RUN curl -sLO https://ffmpeg.org/releases/ffmpeg-4.1.6.tar.bz2 && \
-    tar -jx --strip-components=1 -f ffmpeg-4.1.6.tar.bz2 &&  \
-    ./configure &&  make && \
-    make install
-
-COPY --from=builder /build/main kubeedge/
-COPY ./config.yaml kubeedge/
-
-WORKDIR kubeedge
