+++ staging/src/github.com/kubeedge/mapper-framework/pkg/mapper/data/stream/video.go
